requirements.txt
paho-mqtt
python-dotenv
numpy
pandas
